Update serializers for ngrams api, add tests, fix get_data
