+++ analysis_module/admin.py
@@ -1,3 +1,8 @@
 from django.contrib import admin
 
 # Register your models here.
+
+from .models import AnalysisModuleRequest
+
+
+admin.site.register(AnalysisModuleRequest)
